+++ src/Oro/Bundle/NotificationBundle/Migrations/Schema/v1_0/OroNotificationBundle.php
@@ -51,7 +51,7 @@ class OroNotificationBundle implements Migration
     public static function oroNotificationEmailNotificationTable(Schema $schema, $tableName = null)
     {
         /** Generate table oro_notification_emailnotification **/
-        $table = $schema->createTable($tableName ? : 'oro_notification_emailnotification');
+        $table = $schema->createTable($tableName ?: 'oro_notification_emailnotification');
         $table->addColumn('id', 'integer', ['autoincrement' => true]);
         $table->addColumn('recipient_list_id', 'integer', ['notnull' => false]);
         $table->addColumn('template_id', 'integer', ['notnull' => false]);
@@ -90,9 +90,9 @@ class OroNotificationBundle implements Migration
     public static function oroNotificationRecipientGroupTable(Schema $schema, $tableName = null)
     {
         /** Generate table oro_notification_recipient_group **/
-        $table = $schema->createTable($tableName ? : 'oro_notification_recipient_group');
+        $table = $schema->createTable($tableName ?: 'oro_notification_recipient_group');
         $table->addColumn('recipient_list_id', 'integer', []);
-        $table->addColumn('group_id', 'integer', []);
+        $table->addColumn('group_id', 'smallint', []);
         $table->setPrimaryKey(['recipient_list_id', 'group_id']);
         $table->addIndex(['recipient_list_id'], 'IDX_F6E3D23E2B9E3E89', []);
         $table->addIndex(['group_id'], 'IDX_F6E3D23EFE54D947', []);
@@ -108,7 +108,7 @@ class OroNotificationBundle implements Migration
     public static function oroNotificationRecipientListTable(Schema $schema, $tableName = null)
     {
         /** Generate table oro_notification_recipient_list **/
-        $table = $schema->createTable($tableName ? : 'oro_notification_recipient_list');
+        $table = $schema->createTable($tableName ?: 'oro_notification_recipient_list');
         $table->addColumn('id', 'integer', ['autoincrement' => true]);
         $table->addColumn('email', 'string', ['notnull' => false, 'length' => 255]);
         $table->addColumn('owner', 'boolean', ['notnull' => false]);
@@ -125,7 +125,7 @@ class OroNotificationBundle implements Migration
     public static function oroNotificationRecipientUserTable(Schema $schema, $tableName = null)
     {
         /** Generate table oro_notification_recipient_user **/
-        $table = $schema->createTable($tableName ? : 'oro_notification_recipient_user');
+        $table = $schema->createTable($tableName ?: 'oro_notification_recipient_user');
         $table->addColumn('recipient_list_id', 'integer', []);
         $table->addColumn('user_id', 'integer', []);
         $table->setPrimaryKey(['recipient_list_id', 'user_id']);
@@ -147,9 +147,9 @@ class OroNotificationBundle implements Migration
         $recipientListTableName = null
     ) {
         /** Generate foreign keys for table oro_notification_emailnotification **/
-        $table = $schema->getTable($tableName ? : 'oro_notification_emailnotification');
+        $table = $schema->getTable($tableName ?: 'oro_notification_emailnotification');
         $table->addForeignKeyConstraint(
-            $schema->getTable($recipientListTableName ? : 'oro_notification_recipient_list'),
+            $schema->getTable($recipientListTableName ?: 'oro_notification_recipient_list'),
             ['recipient_list_id'],
             ['id'],
             ['onDelete' => null, 'onUpdate' => null]
@@ -182,7 +182,7 @@ class OroNotificationBundle implements Migration
         $recipientListTableName = null
     ) {
         /** Generate foreign keys for table oro_notification_recipient_group **/
-        $table = $schema->getTable($tableName ? : 'oro_notification_recipient_group');
+        $table = $schema->getTable($tableName ?: 'oro_notification_recipient_group');
         $table->addForeignKeyConstraint(
             $schema->getTable('oro_access_group'),
             ['group_id'],
@@ -190,7 +190,7 @@ class OroNotificationBundle implements Migration
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
         );
         $table->addForeignKeyConstraint(
-            $schema->getTable($recipientListTableName ? : 'oro_notification_recipient_list'),
+            $schema->getTable($recipientListTableName ?: 'oro_notification_recipient_list'),
             ['recipient_list_id'],
             ['id'],
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
@@ -211,7 +211,7 @@ class OroNotificationBundle implements Migration
         $recipientListTableName = null
     ) {
         /** Generate foreign keys for table oro_notification_recipient_user **/
-        $table = $schema->getTable($tableName ? : 'oro_notification_recipient_user');
+        $table = $schema->getTable($tableName ?: 'oro_notification_recipient_user');
         $table->addForeignKeyConstraint(
             $schema->getTable('oro_user'),
             ['user_id'],
@@ -219,7 +219,7 @@ class OroNotificationBundle implements Migration
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
         );
         $table->addForeignKeyConstraint(
-            $schema->getTable($recipientListTableName ? : 'oro_notification_recipient_list'),
+            $schema->getTable($recipientListTableName ?: 'oro_notification_recipient_list'),
             ['recipient_list_id'],
             ['id'],
             ['onDelete' => 'CASCADE', 'onUpdate' => null]
